+++ app/src/main/java/com/example/myapplication/Login/LoginStartFragment.kt
@@ -1,0 +1,22 @@
+package com.example.myapplication.Login
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.myapplication.R
+
+class LoginStartFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_login_start, container, false)
+    }
+
+
+
+}
