+++ app/src/main/res/layout/activity_login.xml
@@ -7,7 +7,6 @@
     android:fitsSystemWindows="true"
     android:theme="@style/AppTheme">
 
-
     <FrameLayout
         android:id="@+id/main_frm"
         android:layout_width="match_parent"
